Fix -Wextra 2/4.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@1200 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/libsrc/wspiapi/WspiapiLegacyGetNameInfo.c
@@ -22,7 +22,7 @@ WspiapiLegacyGetNameInfo(const struct sockaddr *ptSocketAddress,
   char *pszNode = NULL;
   char *pc = NULL;
 
-  if ((!ptSocketAddress) || (tSocketLength < sizeof(struct sockaddr)))
+  if ((!ptSocketAddress) || ((size_t)tSocketLength < sizeof(struct sockaddr)))
     return EAI_FAIL;
   if (ptSocketAddress->sa_family != (unsigned short)AF_INET)
     return EAI_FAMILY;
